feature/refactor: started reworking sya-pyqt options layout
Now the Options window is going to provide seperate element for the
URL and Tracklist (which will be a textbox upcoming next commit).
This is more intuitive for the user than them needing to leave the
application and focus on an external notepad to create the tracklist
and return to sya-pyqt with the path of that.